# Lazy entry points in verdi