--- a/aiida/cmdline/params/types/identifier.py
+++ b/aiida/cmdline/params/types/identifier.py
@@ -10,8 +10,6 @@
 """
 Module for custom click param type identifier
 """
-from __future__ import annotations
-
 from abc import ABC, abstractmethod
 from functools import cached_property
 
@@ -33,7 +31,7 @@ class IdentifierParamType(click.ParamType, ABC):
     which should be a subclass of `aiida.orm.utils.loaders.OrmEntityLoader` for the corresponding orm class.
     """
 
-    def __init__(self, sub_classes: tuple[str] | None = None):
+    def __init__(self, sub_classes=None):
         """
         Construct the parameter type, optionally specifying a tuple of entry points that reference classes
         that should be a sub class of the base orm class of the orm class loader. The classes pointed to by
@@ -52,7 +50,7 @@ class IdentifierParamType(click.ParamType, ABC):
         if sub_classes is not None and not isinstance(sub_classes, tuple):
             raise TypeError('sub_classes should be a tuple of entry point strings')
 
-        self._sub_classes: tuple | None = None
+        self._sub_classes = None
         self._entry_point_strings = sub_classes
 
     @cached_property
@@ -61,7 +59,7 @@ class IdentifierParamType(click.ParamType, ABC):
         from aiida.common import exceptions
 
         if self._entry_point_strings is None:
-            return None
+            return []
 
         entry_points = []
         for entry_point_string in self._entry_point_strings: