… .pre-commit-config.yaml
… aiida/cmdline/params/types/path.py
… aiida/cmdline/utils/__init__.py
… aiida/cmdline/utils/common.py
… aiida/cmdline/utils/echo.py
… aiida/common/lang.py
… aiida/common/timezone.py
… aiida/manage/configuration/__init__.py
… aiida/manage/configuration/config.py
… aiida/manage/configuration/settings.py
… aiida/manage/tests/pytest_fixtures.py
… aiida/orm/groups.py
… aiida/orm/nodes/data/code/abstract.py
… aiida/orm/nodes/data/code/containerized.py
… aiida/orm/nodes/data/code/installed.py
… aiida/orm/nodes/data/code/portable.py
… aiida/orm/nodes/node.py
… aiida/plugins/entry_point.py
… aiida/plugins/factories.py
… aiida/plugins/utils.py
… aiida/transports/cli.py
… docs/source/nitpick-exceptions
… docs/source/reference/command_line.rst
… tests/cmdline/params/types/test_identifier.py
… tests/manage/configuration/test_config.py
… tests/plugins/test_entry_point.py
… tests/test_conftest.py