This is my last PR in a quest for faster import speed, specifically now targetting verdi and its tab-completion.

The main change here is a lazy loading of entry points in verdi commands. This grants us at least 20ms, but usually more since loading entry points is very dependent on HDD cache. The important files to review are `cmdline/params/types/plugin.py` and `cmdline/params/types/identifiers.py`.

There is also an assortment of other lazy imports (including some slow stdlib imports) that together grant sizeable improvement. With these changes the tab completion finally feels almost instant. :tada: 

Further improvements are all in sub-5ms range and would be hard to achieve and maintain.

## New timings

```python
Benchmark 1: python -c "import aiida.cmdline.commands"
  Time (mean ± σ):     120.8 ms ±   1.2 ms    [User: 111.3 ms, System: 9.0 ms]
  Range (min … max):   119.4 ms … 123.7 ms    24 runs
```

```python
Benchmark 1: python -c "import aiida.orm"
  Time (mean ± σ):     310.2 ms ±   5.4 ms    [User: 378.8 ms, System: 256.0 ms]
  Range (min … max):   305.3 ms … 324.6 ms    10 runs
```

```python
Benchmark 1: python -c "import aiida"
  Time (mean ± σ):      54.6 ms ±   0.8 ms    [User: 48.8 ms, System: 3.3 ms]
  Range (min … max):    53.1 ms …  57.8 ms    53 runs
```

## `main`

```python
Benchmark 1: python -c "import aiida.cmdline.commands"
  Time (mean ± σ):     170.8 ms ±   1.5 ms    [User: 152.9 ms, System: 15.5 ms]
  Range (min … max):   168.7 ms … 174.3 ms    17 runs
```

```python
Benchmark 1: python -c "import aiida.orm"
  Time (mean ± σ):     347.9 ms ±   4.1 ms    [User: 405.2 ms, System: 270.4 ms]
  Range (min … max):   344.4 ms … 356.0 ms    10 runs
```

```python
Benchmark 1: python -c "import aiida"
  Time (mean ± σ):      93.0 ms ±   0.9 ms    [User: 84.6 ms, System: 7.0 ms]
  Range (min … max):    91.7 ms …  95.1 ms    31 runs
```

TODO: Would be good to extend some test to ensure that entry points are not loaded during shell completion.