@danielhollas 2023-10-20
> Technically it can also be `None` so this would except. But then again the original `pathlib.Path(None)` would also except. Guess this is a potential bug lying in the wait. Fine to keep this change for now

Yeah, that's a good point. In practice the `set_configuration_directory()`, which is called on import, ensures that it is not None. But we might still be worried by the variable being overridden from outside so I decided to assign the defaults at the top of the file, and enable type checking which should give us more confidence. Please take a look.

I've also added/enabled type checking for plugin.py and identifier.py so I have more confidence I did not break anything. (I was surprised to discover the huge list of mypy-excluded files in pre-commit config  🙈  😅)