@sphuber 2023-10-21
> I've also added/enabled type checking for plugin.py and identifier.py so I have more confidence I did not break anything. (I was surprised to discover the huge list of mypy-excluded files in pre-commit config 🙈 😅)

Well, this is because we only recently (since a year or so) started adding typing. Originally we configured `mypy` to not require any typing and have a very lax configuration. This we changed at some point to be very strict and instead explicitly exclude things that weren't typed yet. This to enforce that newly added code would have to be typed. In the meantime, we would have to work to slowly remove existing files from the exclude list, but that is a lot of work of course :D Thanks for pitching in!