@danielhollas 2023-10-22
Thank you for fixing the docs! I honestly can't even find the errors in the damn readthedocs build log. :see_no_evil: 

> Well, this is because we only recently (since a year or so) started adding typing. Originally we configured mypy to not require any typing and have a very lax configuration. This we changed at some point to be very strict and instead explicitly exclude things that weren't typed yet. This to enforce that newly added code would have to be typed.

Yeah, that makes total sense. It was more about me being very confused why mypy was ignoring these files, since I was only looking at the configuration in `pyproject.toml`. It did not occur to me that it would be configured in pre-commit.