@sphuber 2023-10-22
> Thank you for fixing the docs! I honestly can't even find the errors in the damn readthedocs build log

yeah, they are pretty hidden if you don't know what to look for. Typically, the errors start after the `writing output...` lines and just before `highlighting module code...` The warnings typically start with the path of the relevant file, e.g. `/home/docs/checkouts/readthedocs.org/user_builds/aiida-core/envs/6153/lib/python3.10/`. They are pretty long and with the small windowed box of RTD you always have to scroll to the right to them, but there you will see something like
```
WARNING: py:class reference target not found: EntryPoints
```